requirements.txt
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.1
Flask==2.0.1
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.8.6
PyYAML==6.0
requests==2.27.1
SQLAlchemy==1.4.18
urllib3==1.26.8
Werkzeug==2.0.1
